Discord Bot/requirements.txt
discord.py>=2.4.0
aiohttp>=3.10.10
aiosqlite>=0.20.0
python-dotenv>=1.0.1
PyNaCl>=1.5.0,<1.6
audioop-lts>=0.2.2
discord-ext-voice_recv @ git+https://github.com/imayhaveborkedit/discord-ext-voice-recv.git@ac04ea7b0941112e83767cf1c1469b408fa06748
faster-whisper>=1.2.1
google-genai>=1.64.0